refactor: have separate error modules in each laye
