+++ src/app/errors/token_error.rs
