+++ src/app/token.rs
@@ -2,19 +2,19 @@ use std::fs::{self, File};
 use std::io::Write;
 use std::path::PathBuf;
 
-use crate::errors::token_error::TokenError;
-use crate::errors::CliError;
+use super::errors::token_error::TokenError;
+use super::errors::AppError;
 
 pub const TOKEN_FILE_NAME: &str = "GALOY_TOKEN";
 pub const TOKEN_FOLDER_NAME: &str = ".galoy-cli";
 
-pub fn get_token_file_path() -> Result<PathBuf, CliError> {
+pub fn get_token_file_path() -> Result<PathBuf, AppError> {
     let home_dir = dirs::home_dir().ok_or(TokenError::HomeDirectoryNotFound)?;
     let token_dir = home_dir.join(TOKEN_FOLDER_NAME);
     Ok(token_dir.join(TOKEN_FILE_NAME))
 }
 
-pub fn get_token() -> anyhow::Result<Option<String>> {
+pub fn get_token() -> Result<Option<String>, AppError> {
     let token_file = get_token_file_path()?;
 
     if token_file.exists() {
@@ -26,7 +26,7 @@ pub fn get_token() -> anyhow::Result<Option<String>> {
     }
 }
 
-pub fn save_token(token: &str) -> Result<(), CliError> {
+pub fn save_token(token: &str) -> Result<(), AppError> {
     let token_file = get_token_file_path()?;
     let parent_dir = token_file.parent().unwrap().to_owned();
     fs::create_dir_all(&parent_dir)
@@ -41,13 +41,13 @@ pub fn save_token(token: &str) -> Result<(), CliError> {
     Ok(())
 }
 
-pub fn remove_token() -> Result<(), CliError> {
+pub fn remove_token() -> Result<(), AppError> {
     let token_file = get_token_file_path()?;
     if token_file.exists() {
         fs::remove_file(&token_file)
             .map_err(|_| TokenError::FailedToDeleteFile(token_file.clone()))?;
         Ok(())
     } else {
-        Err(CliError::TokenError(TokenError::TokenFileNotFound))
+        Err(AppError::TokenError(TokenError::TokenFileNotFound))
     }
 }
