+++ src/client/errors/api_error.rs
