+++ src/client/errors/me_error.rs
