+++ src/client/errors/mod.rs
@@ -1,17 +1,14 @@
 use thiserror::Error;
 
-use self::{api_error::ApiError, me_error::MeError, token_error::TokenError};
+use self::{api_error::ApiError, me_error::MeError};
 
 pub mod api_error;
 pub mod me_error;
-pub mod token_error;
 
 #[derive(Error, Debug)]
-pub enum CliError {
+pub enum ClientError {
     #[error(transparent)]
     MeError(#[from] MeError),
     #[error(transparent)]
     ApiError(#[from] ApiError),
-    #[error(transparent)]
-    TokenError(#[from] TokenError),
 }
