+++ src/client/requests/me.rs
@@ -1,12 +1,12 @@
 use crate::client::{
+    errors::{api_error::ApiError, me_error::MeError, ClientError},
     queries::{query_me, QueryMe, QueryMeMe},
     GaloyClient,
 };
-use crate::errors::{api_error::ApiError, me_error::MeError, CliError};
 use graphql_client::reqwest::post_graphql;
 
 impl GaloyClient {
-    pub async fn me(&self) -> Result<QueryMeMe, CliError> {
+    pub async fn me(&self) -> Result<QueryMeMe, ClientError> {
         let variables = query_me::Variables;
 
         let response_body = post_graphql::<QueryMe, _>(&self.graphql_client, &self.api, variables)
