+++ src/lib.rs
@@ -1,5 +1,4 @@
 mod app;
 mod client;
-mod errors;
 
 pub mod cli;
